requirements.txt
google-adk==1.28.0
google-generativeai==0.8.5
flask==3.1.3
python-dotenv==1.2.1
flask-cors==6.0.2
